pyproject.toml
[project]
name = "tagmemo-py"
version = "0.1.0"
description = "TagMemo Super RAG — 基于浪潮算法的智能记忆对话系统 (Python 版)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["rag", "tagmemo", "memory", "ai", "embedding"]

dependencies = [
    # === Web 框架 ===
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # === 向量索引 (替代 rust-vexus-lite) ===
    "usearch>=2.12.0",
    # === 数值计算 (替代 Float32Array + nalgebra) ===
    "numpy>=1.26.0",
    # === HTTP 客户端 (替代 axios / node-fetch) ===
    "httpx>=0.27.0",
    # === Token 计数 (替代 @dqbd/tiktoken) ===
    "tiktoken>=0.7.0",
    # === HTML 解析 (替代 cheerio) ===
    "beautifulsoup4>=4.12.0",
    # === 文件监听 (替代 chokidar) ===
    "watchdog>=4.0.0",
    # === 环境变量 ===
    "python-dotenv>=1.0.0",
    # === 时间处理 ===
    "python-dateutil>=2.9.0",
    # === Windows 时区数据 (zoneinfo 需要) ===
    "tzdata>=2024.1; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
tagmemo = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tagmemo"]

# --------------- ruff ---------------
[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

# --------------- pytest ---------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
